Chips on the Bar are Now Complete
- Fixed a lot of bugs regarding animation

- Added hovering, unhovering, and  purchase animations to the chips

- Owned chips will now be shown on the right of the bar while purchasable chips will be on the left

- Removed the bankruptcy feature since it didn't align well with the overall vision of the game

- Further stability improvements and edge cases safety nets to ensure consistency

The chips initial set up is now fully completed and polished, next up is tying it all together with the other game modes (using chips to play instead of just a text box to enter the amount of money). I will also revisit the bar later on to add more features like drinking. Other than that, there's only the level design and the spin to win wheel to be completed and the game should be ready to be released.